requirements.txt
construct
pysam
pytest
pytest-console-scripts
tox
google-cloud-storage
numpy
pandas
tinydb
firecloud
xmltodict
argparse
